Migrate from Railway/MariaDB to Render/PostgreSQL
- Replace railway.toml with render.yaml (free web service + free Postgres DB)
- Swap MySqlConnector for Npgsql; rename MariaDbConnectionFactory → PostgresConnectionFactory
- Convert all 12 SQL migrations from MySQL to PostgreSQL syntax
  (SERIAL PKs, TIMESTAMPTZ, BOOLEAN, JSONB, separate CREATE INDEX statements)
- Port all MySQL-specific SQL in repositories to PostgreSQL equivalents:
  LAST_INSERT_ID() → RETURNING Id
  INSERT IGNORE → ON CONFLICT DO NOTHING
  ON DUPLICATE KEY UPDATE → ON CONFLICT (id) DO UPDATE SET
  UTC_TIMESTAMP(3) → NOW()
  IsActive = 1/0 → TRUE/FALSE
- Fix sbyte → short for SMALLINT columns (ActivityRow.EventType, CardRow.Priority)
- Update appsettings.json connection string key/format for Npgsql

+++ src/TaskOrchestrator.Api/Persistence/Repositories/ActivityRepository.cs
@@ -46,18 +46,18 @@ public sealed class ActivityRepository(IDbConnectionFactory db) : IActivityRepos
                   UserId = userId, DisplayName = displayName, Description = description });
     }
 
-    // MariaDB TINYINT projects as sbyte; convert explicitly to enum-backed DTO.
+    // PostgreSQL SMALLINT projects as short; convert explicitly to enum-backed DTO.
     private sealed record ActivityRow(
         int Id,
         int CardId,
         int BoardId,
-        sbyte EventType,
+        short EventType,
         string? UserId,
         string? UserDisplayName,
         string Description,
         DateTime OccurredAtUtc)
     {
         public ActivityEventDto ToDto() =>
-            new(Id, CardId, BoardId, (ActivityEventType)EventType, UserId, UserDisplayName, Description, OccurredAtUtc);
+            new(Id, CardId, BoardId, (ActivityEventType)(int)EventType, UserId, UserDisplayName, Description, OccurredAtUtc);
     }
 }
